+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+---------------
+##### Fixed
+- Setvalue/xforms-value-changed inside a select multiple question is not working (enketo-transformer).
+
 [5.17.1] - 2020-01-18
 ----------------------
 ##### Changed